requirements.txt
-i https://pypi.org/simple
amply==0.1.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
apiflask==1.2.0
apispec==6.0.2 ; python_version >= '3.7'
asttokens==2.2.1
authlib==1.2.0
backcall==0.2.0
blinker==1.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.1.3 ; python_version >= '3.7'
contourpy==1.0.7 ; python_version >= '3.8'
cycler==0.11.0 ; python_version >= '3.6'
decorator==5.1.1 ; python_version >= '3.5'
discretemarkovchain==0.22
docutils==0.19 ; python_version >= '3.7'
executing==1.2.0
faker==13.3.4
faker_education
flask==2.1.0
flask-cors==3.0.10
flask-httpauth==4.7.0
flask-mail==0.9.1
flask-marshmallow==0.14.0
flask-sqlalchemy==2.5.1
fonttools==4.38.0 ; python_version >= '3.7'
greenlet==2.0.2 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
ipython==8.10.0 ; python_version >= '3.8'
itsdangerous==2.1.2 ; python_version >= '3.7'
jedi==0.18.2 ; python_version >= '3.6'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
kiwisolver==1.4.4 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
marshmallow==3.19.0 ; python_version >= '3.7'
nanoid==2.0.0
networkx==3.1
numpy==1.24.2 ; python_version >= '3.8'
packaging==23.0 ; python_version >= '3.7'
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.4.0 ; python_version >= '3.7'
prompt-toolkit==3.0.36 ; python_full_version >= '3.6.2'
ptyprocess==0.7.0
pulp==2.3
pure-eval==0.2.2
py-rankaggregation==0.1.2
pygments==2.14.0 ; python_version >= '3.6'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
pyrebase4==4.7.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
pytz==2022.7.1
requests==2.29.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.46
stack-data==0.6.2
threadpoolctl==3.1.0 ; python_version >= '3.6'
traitlets==5.9.0 ; python_version >= '3.7'
wcwidth==0.2.6
webargs==8.2.0 ; python_full_version >= '3.7.2'
werkzeug==2.2.3 ; python_version >= '3.7'
flask
python-dotenv
flask-sqlalchemy
psycopg2-binary==2.8.6
